% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFilesS3.R
\name{bucket_contents_to_shiny_files_df}
\alias{bucket_contents_to_shiny_files_df}
\title{Title}
\usage{
bucket_contents_to_shiny_files_df(bucket_contents)
}
\arguments{
\item{bucket_contents}{contents of a bucket from s3tools}
}
\value{
df containing everything required for shiny files
}
\description{
Title
}
\examples{
bucket_contents_to_shiny_files_df(bucket_contents)
}
