% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_browser.R
\name{file_explorer_s3}
\alias{file_explorer_s3}
\title{s3 file explorer}
\usage{
file_explorer_s3()
}
\description{
Selected files will be inseted as s3tools commands into the source pane
}
\details{
Browser currently supports selection of multiple files.
}
