% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFilesS3.R
\name{s3_dir_shiny_files}
\alias{s3_dir_shiny_files}
\title{Title}
\usage{
s3_dir_shiny_files(current_path = "")
}
\arguments{
\item{current_path}{}
}
\value{
df with required shinyfiles fields
}
\description{
Title
}
\examples{
s3_dir_shiny_files('bucket/dir/')
}
