% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFilesS3.R
\name{s3_dir_shiny_files_list}
\alias{s3_dir_shiny_files_list}
\title{Title}
\usage{
s3_dir_shiny_files_list(roots, restrictions, filetypes, hidden = FALSE)
}
\arguments{
\item{hidden}{}
}
\value{
a list compatable with shinyfiles
}
\description{
Title
}
