% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFilesS3.R
\name{s3time_to_shinyfiles}
\alias{s3time_to_shinyfiles}
\title{s3time to shinyfiles time convertor}
\usage{
s3time_to_shinyfiles(time)
}
\arguments{
\item{string}{containing s3 time}
}
\value{
shinyfile compatible time strings
}
\description{
s3time to shinyfiles time convertor
}
\examples{
s3time_to_shinyfiles(s3times)
}
