% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFileParsers.R
\name{shiy_js_to_s3tools_command}
\alias{shiy_js_to_s3tools_command}
\title{Title}
\usage{
shiy_js_to_s3tools_command(df_in)
}
\arguments{
\item{df_in}{}
}
\value{
formatted s3 path command
}
\description{
Title
}
